[refact] Move around some of the functionality
