+++ project1/feature/timer/Timer.jsx
@@ -1,12 +1,29 @@
 import React from 'react'
 import { StyleSheet, Text, View } from 'react-native'
 
-import { convertToDisplayTime } from '../utils'
+import { convertToDisplayTime, vibrate } from '../../utils'
+
+const styles = StyleSheet.create({
+  timerView: {
+    alignItems: 'center',
+  },
+  timerText: {
+    fontSize: 40,
+  },
+})
 
 const Timer = ({ startWorkTime, startBreakTime }) => {
   const [useWorkTime, setUseWorkTime] = React.useState(true)
   const [time, setTime] = React.useState(startWorkTime)
   let interval = null
+
+  const resetTime = React.useCallback(() => {
+    if (useWorkTime) {
+      setTime(startBreakTime)
+    } else {
+      setTime(startWorkTime)
+    }
+  }, [useWorkTime])
 
   React.useEffect(() => {
     interval = setInterval(() => {
@@ -20,20 +37,16 @@ const Timer = ({ startWorkTime, startBreakTime }) => {
 
   React.useEffect(() => {
     if (time < 0) {
-      if (useWorkTime) {
-        setTime(startBreakTime)
-      } else {
-        setTime(startWorkTime)
-      }
-
+      // vibrate()
+      resetTime()
       setUseWorkTime(!useWorkTime)
     }
   }, [time])
 
   return (
-    <View>
-      <Text>{useWorkTime ? 'Work' : 'Relax'}</Text>
-      <Text>Timer: {convertToDisplayTime(time)}</Text>
+    <View style={styles.timerView}>
+      <Text style={styles.timerText}>{useWorkTime ? 'Work' : 'Relax'}</Text>
+      <Text style={styles.timerText}>Timer: {convertToDisplayTime(time)}</Text>
     </View>
   )
 }
